Refactoring + added editable comments to private annotations
