+++ src/components/Annotation/Comment/CommentProps.tsx
@@ -15,8 +15,6 @@ export interface CommentProps {
 
   present: PresentUser[];
 
-  onUpdateComment(comment: AnnotationBody, updated: AnnotationBody): void;
-
   onDeleteAnnotation(): void;
   
 }
