+++ src/components/Annotation/Comment/PrivateComment/PrivateCommentActions.tsx
@@ -1,5 +1,5 @@
 import * as Dropdown from '@radix-ui/react-dropdown-menu';
-import { DotsThreeVertical, Trash, UsersThree } from '@phosphor-icons/react';
+import { DotsThreeVertical, Pencil, Trash, UsersThree } from '@phosphor-icons/react';
 import type { Translations } from 'src/Types';
 
 const { Content, Item, Root, Trigger } = Dropdown;
@@ -11,6 +11,8 @@ interface PrivateCommentActionsProps {
   isFirst?: boolean;
 
   onMakePublic(): void;
+
+  onEditComment(): void;
 
   onDeleteAnnotation(): void;
 
@@ -35,17 +37,21 @@ export const PrivateCommentActions = (props: PrivateCommentActionsProps) => {
           {props.isFirst && (
             <>
               <Item className="dropdown-item" onSelect={props.onMakePublic}>
-                <UsersThree size={16} /> <span>{t['Make this annotation public']}</span>
+                <UsersThree size={16} /> <span>{t['Make annotation public']}</span>
               </Item>
 
               <Item className="dropdown-item" onSelect={props.onDeleteAnnotation}>
-                <Trash size={16} /> <span>{t['Delete this annotation']}</span>
+                <Trash size={16} /> <span>{t['Delete annotation']}</span>
               </Item>
             </>
           )}
 
+          <Item className="dropdown-item" onSelect={props.onEditComment}>
+            <Pencil size={16} /> <span>{t['Edit comment']}</span>
+          </Item>
+
           <Item className="dropdown-item" onSelect={props.onDeleteComment}>
-            <Trash size={16} /> <span>{t['Delete this comment']}</span>
+            <Trash size={16} /> <span>{t['Delete comment']}</span>
           </Item>
         </div>
       </Content>
