+++ src/components/Annotation/Comment/PublicComment/PublicComment.tsx
@@ -1,9 +1,9 @@
-import { useEffect, useRef, useState } from 'react';
-import TextareaAutosize from 'react-textarea-autosize';
+import { useState } from 'react';
 import { useAnnotationStore, useAnnotatorUser } from '@annotorious/react';
 import type { PresentUser, User } from '@annotorious/react';
 import { Creator } from '../../Creator';
 import type { CommentProps } from '../CommentProps';
+import { EditableComment } from '../EditableComment';
 import { PublicCommentActions } from './PublicCommentActions';
 
 import '../Comment.css';
@@ -12,53 +12,16 @@ export const PublicComment = (props: CommentProps) => {
 
   const { comment, present } = props;
 
-  const textarea = useRef<HTMLTextAreaElement>(null);
-
-  const [editable, setEditable] = useState(false);
-
-  const [value, setValue] = useState(comment.value);
-
   const me = useAnnotatorUser();
 
   const store = useAnnotationStore();
+
+  const [editable, setEditable] = useState(false);
 
   const creator: PresentUser | User | undefined = 
     present.find(p => p.id === comment.creator?.id) || comment.creator;
 
   const isMine = creator?.id === me.id;
-
-  useEffect(() => {
-    const { current } = textarea;
-
-    if (editable && current) {
-      // Put this in the event queue, so that 
-      // Radix trigger focus happens first, textarea
-      // focus second
-      setTimeout(() => {
-        current?.focus({ preventScroll: true });
-
-        // This trick sets the cursor to the end of the text
-        current.value = '';
-        current.value = comment.value;
-      }, 1);
-    }
-  }, [editable]);
-
-  const onSaveChange = (evt: React.FormEvent) => {
-    evt.preventDefault();
-
-    props.onUpdateComment(comment, {
-      ...comment,
-      value: textarea.current!.value
-    });
-
-    setEditable(false);
-  }
-
-  const onCancelChange = () => {
-    setEditable(false);
-    setValue(comment.value);
-  }
 
   const onDeleteComment = () => store.deleteBody(comment);
 
@@ -73,31 +36,12 @@ export const PublicComment = (props: CommentProps) => {
         creator={creator} 
         createdAt={comment.created} />
 
-      {editable ? (
-        <form onSubmit={onSaveChange}>
-          <TextareaAutosize 
-            className="no-drag"
-            ref={textarea}
-            value={value}
-            onChange={evt => setValue(evt.target.value)}
-            rows={1} 
-            maxRows={10} />
-
-          <div className="buttons">
-            <button 
-              disabled={value === comment.value}
-              className="primary sm flat"
-              type="submit">Save</button>
-
-            <button 
-              className="sm flat"
-              type="button"
-              onClick={onCancelChange}>Cancel</button>
-          </div>
-        </form>
-      ) : (
-        <p className="no-drag">{comment.value}</p>
-      )}
+      <EditableComment 
+        editable={editable}
+        i18n={props.i18n} 
+        comment={comment} 
+        onChanged={() => setEditable(false)} 
+        onCanceled={() => setEditable(false)} />
 
       {isMine && (
         <PublicCommentActions
