+++ src/i18n/de/annotation-common.json
@@ -14,5 +14,5 @@
   "Edit comment": "Kommentar bearbeiten",
   "Delete annotation": "Annotation löschen",
   "Delete comment": "Kommentar löschen",
-  "Make this annotation public": "Diese Annotation öffentlich machen"
+  "Make annotation public": "Diese Annotation öffentlich machen"
 }
