+++ src/i18n/en/annotation-common.json
@@ -14,5 +14,5 @@
   "Edit comment": "Edit comment",
   "Delete annotation": "Delete annotation",
   "Delete comment": "Delete comment",
-  "Make this annotation public": "Make this annotation public"
+  "Make annotation public": "Make this annotation public"
 }
